+++ devcovenant/builtin/profiles/swift/assets/Package.swift
@@ -1,3 +1,3 @@
 // swift-tools-version:5.9
 import PackageDescription
-let package = Package(name: "devcovenant")
+let package = Package(name: "project-name")